+++ src/allotropy/to_allotrope.py
@@ -1,12 +1,14 @@
 from datetime import tzinfo
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any, Optional, Union
 
 from allotropy.allotrope.allotrope import serialize_and_validate_allotrope
 from allotropy.exceptions import AllotropeConversionError
 from allotropy.named_file_contents import NamedFileContents
-from allotropy.parser_factory import get_parser, VendorType
+from allotropy.parser_factory import get_parser, Vendor
 from allotropy.types import IOType
+
+VendorType = Union[Vendor, str]
 
 
 def allotrope_from_io(
@@ -25,8 +27,13 @@ def allotrope_model_from_io(
     vendor_type: VendorType,
     default_timezone: Optional[tzinfo] = None,
 ) -> Any:
+    try:
+        vendor = Vendor(vendor_type)
+    except ValueError as e:
+        error = f"Failed to create parser, unregistered vendor: {vendor_type}."
+        raise AllotropeConversionError(error) from e
     named_file_contents = NamedFileContents(contents, filename)
-    parser = get_parser(vendor_type, default_timezone=default_timezone)
+    parser = get_parser(vendor, default_timezone=default_timezone)
     return parser.to_allotrope(named_file_contents)